don't run more than one ffmpeg at a time; always remove shrunk video

+++ naturewatch_camera_server/Publisher.py
@@ -3,6 +3,7 @@ from subprocess import call
 import logging
 import threading
 import os
+import os.path
 
 class Publisher():
 
@@ -14,6 +15,12 @@ class Publisher():
         else:
             self.logger = logging
 
+        # The semaphore number sets how many publishing tasks to do at the same time.
+        # Since FFmpeg is multi-threaded, there's no benefit in running more than one at the same time.
+        # Also the high load can make the detection logic fail.
+        # Increase it at your own risk.
+        self.ffmpegSemaphore = threading.Semaphore(1)
+
         self.updater = Updater(self.config["telegram_api_key"])
         # dp = updater.dispatcher
         self.logger.info("Starting Telegram publisher")
@@ -21,36 +28,39 @@ class Publisher():
         self.updater.start_polling()
         self.logger.info("Telegram publisher started")
 
-    def shrinkAndSend(self, video_file_name, thumb_file_name):
-        self.logger.info("Shrinking video for publishing: " + video_file_name)
-        
-        shrunk_file_name = video_file_name.replace(".mp4", "_shrunk.mp4")
-        call(["ffmpeg", 
-            "-hide_banner",
-            "-nostats",
-            "-i", video_file_name, 
-            "-vf", "scale=iw/2:ih/2", 
-            "-crf" , "25", 
-            shrunk_file_name])
+    def doPublish(self, video_file_name, thumb_file_name):
+        with self.ffmpegSemaphore:
+            self.logger.info("Shrinking video for publishing: " + video_file_name)
+            
+            shrunk_file_name = video_file_name.replace(".mp4", "_shrunk.mp4")
+            try:
+                call(["ffmpeg", 
+                    "-hide_banner",
+                    "-nostats",
+                    "-i", video_file_name, 
+                    "-vf", "scale=960:540", 
+                    "-crf" , "25", 
+                    shrunk_file_name])
 
-        with open(shrunk_file_name, 'rb') as video_file:
-            with open(thumb_file_name, 'rb') as thumbnail_file:
-                self.logger.info("sending video file")
-                self.updater.bot.send_video(
-                    chat_id=self.config['telegram_chat_id'],
-                    video=video_file,
-                    thumb=thumbnail_file,
-                    width=1920,
-                    height=1080,
-                    supports_streaming=True)
-                self.logger.info("send completed")
-        
-        os.remove(shrunk_file_name)
+                with open(shrunk_file_name, 'rb') as video_file:
+                    with open(thumb_file_name, 'rb') as thumbnail_file:
+                        self.logger.info("sending video file")
+                        self.updater.bot.send_video(
+                            chat_id=self.config['telegram_chat_id'],
+                            video=video_file,
+                            thumb=thumbnail_file,
+                            width=960,
+                            height=540,
+                            supports_streaming=True)
+                        self.logger.info("send completed")
+            finally:
+                if os.path.isfile(shrunk_file_name):      
+                    os.remove(shrunk_file_name)
 
     def publish_image(self, file_name):
         pass
 
     def publish_video(self, video_file_name, thumb_file_name):
-        thread = threading.Thread(target=self.shrinkAndSend, args=(video_file_name, thumb_file_name))
-        thread.start()
+        thread = threading.Thread(target=self.doPublish, args=(video_file_name, thumb_file_name))
         self.logger.info("Will publish using thread " + str(thread))
+        thread.start()
